JS can read AND write!
connecting button still works.

hello world button sends "hello" in an encrypted form to arduino, which (when echo is running) can repeat it back. this repeated phrase is sent back to the website, which decrypts it and prints it on the console!

problem right now is that the button only works once, then returns an error